requirements.txt
ArchivesSnake==0.3.0
attrs==18.1.0
boltons==18.0.0
certifi==2018.4.16
chardet==3.0.4
clinner==1.9.4
colorama==0.3.9
colorlog==3.1.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.0
django-common-helpers==0.9.1
django-cron==0.5.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
ElectronBonder==0.2.0
gitdb2==2.0.3
GitPython==2.1.9
health-check==3.4.1
idna==2.6
itypes==1.1.0
Jinja2==2.10
jq==0.1.6
MarkupSafe==1.0
psycopg2-binary==2.7.4
PyJWT==1.6.1
pytz==2018.4
PyYAML==3.12
requests==2.18.4
six==1.11.0
smmap2==2.0.3
structlog==18.1.0
uritemplate==3.0.0
urllib3==1.22
uuid==1.30
